Fix bug to support multple subgraphs extending entities
