+++ src/test/groovy/com/intuit/graphql/orchestrator/integration/federation/EntityTypeSpec.groovy
@@ -1,21 +1,16 @@
-package com.intuit.graphql.orchestrator.integration
+package com.intuit.graphql.orchestrator.integration.federation
 
-import com.google.common.collect.ImmutableMap
+
 import com.intuit.graphql.orchestrator.ServiceProvider
 import com.intuit.graphql.orchestrator.TestHelper
 import com.intuit.graphql.orchestrator.TestServiceProvider
 import com.intuit.graphql.orchestrator.schema.Operation
-import com.intuit.graphql.orchestrator.schema.RuntimeGraph
-import graphql.ExecutionInput
-import graphql.ExecutionResult
 import graphql.schema.GraphQLObjectType
 import graphql.schema.GraphQLSchema
 import helpers.BaseIntegrationTestSpecification
 import spock.lang.Subject
 
-import static org.assertj.core.api.Assertions.assertThat
-
-class StitchingTopLevelSpec extends BaseIntegrationTestSpecification {
+class EntityTypeSpec extends BaseIntegrationTestSpecification {
     ServiceProvider employeeProvider, inventoryProvider, reviewsProvider
 
     @Subject
@@ -68,51 +63,6 @@ class StitchingTopLevelSpec extends BaseIntegrationTestSpecification {
         queryType.getFieldDefinition("getStoreByIdAndName") != null
     }
 
-    def "Federation shared value types do not conflict"(){
-        given:
-        def schema1 = """
-                    type Query {
-                        getProvider1Val: sharedValueType
-                    }
-                    type sharedValueType {
-                        id: ID!
-                        name: String
-                    }
-                """
-
-        def schema2 = """
-                    type Query {
-                        getProvider2Val: sharedValueType
-                    }
-                    type sharedValueType {
-                        id: ID!
-                        name: String
-                        test: String
-                    }
-                """
-
-        def valueProvider1 = TestServiceProvider.newBuilder()
-                .namespace("A")
-                .serviceType(ServiceProvider.ServiceType.FEDERATION_SUBGRAPH)
-                .sdlFiles(ImmutableMap.of("schema1", schema1))
-                .build()
-
-        def valueProvider2 = TestServiceProvider.newBuilder()
-                .namespace("B")
-                .serviceType(ServiceProvider.ServiceType.FEDERATION_SUBGRAPH)
-                .sdlFiles(ImmutableMap.of("schema2", schema2))
-                .build()
-
-        when:
-        specUnderTest = createGraphQLOrchestrator([valueProvider1, valueProvider2])
-
-        then:
-        final GraphQLObjectType queryType = specUnderTest?.runtimeGraph?.getOperation(Operation.QUERY)
-        queryType.getFieldDefinition("getProvider1Val").type.name == "sharedValueType"
-        queryType.getFieldDefinition("getProvider2Val").type.name == "sharedValueType"
-        //todo verify the new field in federation spec
-    }
-
     def "Federation entities can be extended with extends directive and extend keyword"() {
         when:
         specUnderTest = createGraphQLOrchestrator([employeeProvider, inventoryProvider, reviewsProvider])
@@ -121,5 +71,19 @@ class StitchingTopLevelSpec extends BaseIntegrationTestSpecification {
         final GraphQLSchema graphQLSchema = specUnderTest?.runtimeGraph?.getExecutableSchema()
         graphQLSchema.getType("Store")?.getFieldDefinition("review") != null
         graphQLSchema.getType("Employee")?.getFieldDefinition("review") != null
+    }
+
+    def "Federation multiple providers can extend entities"() {
+        when:
+        specUnderTest = createGraphQLOrchestrator([employeeProvider, inventoryProvider, reviewsProvider])
+
+        then:
+        final GraphQLSchema graphQLSchema = specUnderTest?.runtimeGraph?.getExecutableSchema()
+
+        graphQLSchema.getType("Employee")?.getFieldDefinition("id") != null
+        graphQLSchema.getType("Employee")?.getFieldDefinition("username") != null
+        graphQLSchema.getType("Employee")?.getFieldDefinition("password") != null
+        graphQLSchema.getType("Employee")?.getFieldDefinition("review") != null
+        graphQLSchema.getType("Employee")?.getFieldDefinition("favoriteItem") != null
     }
 }
