+++ src/test/resources/top_level/federation/review.graphqls
@@ -12,4 +12,5 @@ type Review {
     reviewId: ID!
     rating: Int!
     comments: String
+    test: String
 }
